Restructure project packages
