+++ io/graph_io.py
@@ -1,7 +1,7 @@
 
 import json
 from core.models import GraphData
-from serializers import graph_to_dict, dict_to_graph
+from .serializers import graph_to_dict, dict_to_graph
 
 def export_graph_to_json(graph: GraphData, path: str):
     """Exporte un graphique (et ses courbes) vers un fichier JSON."""
