+++ ui/dialogs/curve_selection_dialog.py
