+++ ui/dialogs/import_curve_dialog.py
