+++ ui/layout_manager.py
