+++ ui/widgets/satellite_item.py
